refactor(inbox): extract the detection property table (#994)
Moves the detectionProps PropertyDef builder out of InboxDetail into
inboxDetectionProps.ts. Pure data construction with no component state,
so it lifts out whole; InboxDetail now calls it with the values it
already had in scope.

InboxDetail.tsx: 1026 -> 930 lines.
